Exception tests

+++ libraries/chain/include/koinos/chain/thunk_utils.hpp
@@ -123,7 +123,7 @@ KOINOS_TODO( "Invoke smart contract sys call handler" )
             [&]( auto& _a ) {                                                                                        \
                KOINOS_THROW( unknown_system_call,                                                                    \
                   "system call table dispatch entry ${sid} has unimplemented type ${tag}",                           \
-                  ("sid", _sid)("tag", (uint64_t)_target.index()) );                                                 \
+                  ("sid", _sid)("tag", _target.index()) );                                                           \
             } }, _target );                                                                                          \
       BOOST_PP_IF(_THUNK_IS_VOID(RETURN_TYPE),,return _ret;)                                                         \
    }                                                                                                                 \
